Fix bugs in darjeeling

+++ src/lib/wkpf/include/common/wkpf_properties.h
@@ -48,8 +48,7 @@ extern uint8_t wkpf_write_property_refresh_rate(wuobject_t *wuobject, uint8_t pr
 extern uint8_t wkpf_get_property_status(wuobject_t *wuobject, uint8_t property_number, uint8_t *status);
 
 extern uint8_t wkpf_property_needs_initialisation_push(wuobject_t *wuobject, uint8_t property_number);
-
-extern uint8_t wkpf_update_map_in_flash(uint16_t component_id, uint16_t orig_node_id, uint8_t orig_port_number, uint16_t new_node_id, uint8_t new_port_number);
+extern uint8_t wkpf_update_map_in_flash(uint16_t component_id, uint32_t orig_node_id, uint8_t orig_port_number, uint32_t new_node_id, uint8_t new_port_number);
 extern uint8_t wkpf_update_link_in_flash(uint8_t* orig_link, uint8_t* new_link);
 
 extern bool wkpf_property_status_is_dirty(uint8_t status);
